infra/recognition/pyproject.toml
[project]
name = "recognition"
version = "0.2.0"
description = "InsightFace Recognition Service - SabaiFace API Compatible"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "fastapi>=0.128.0",
    "python-multipart>=0.0.21",
    "httpx>=0.28.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
recognition = "main:app"

# Run commands:
# uv run uvicorn main:app --reload --port 8087  (dev)
# uv run python main.py  (start)
